refactor: changing files structure to better reading and single import
