+++ lib/src/widgets/focus_menu_holder.dart
@@ -1,0 +1,108 @@
+import 'package:flutter/material.dart';
+import 'package:focused_menu/src/models/focused_menu_item.dart';
+import 'package:focused_menu/src/widgets/focused_menu_datails.dart';
+
+class FocusedMenuHolder extends StatefulWidget {
+  final Widget child;
+  final double? menuItemExtent;
+  final double? menuWidth;
+  final List<FocusedMenuItem> menuItems;
+  final bool? animateMenuItems;
+  final BoxDecoration? menuBoxDecoration;
+  final Function onPressed;
+  final Duration? duration;
+  final double? blurSize;
+  final Color? blurBackgroundColor;
+  final double? bottomOffsetHeight;
+  final double? menuOffset;
+
+  /// Open with tap insted of long press.
+  final bool openWithTap;
+
+  const FocusedMenuHolder(
+      {Key? key,
+      required this.child,
+      required this.onPressed,
+      required this.menuItems,
+      this.duration,
+      this.menuBoxDecoration,
+      this.menuItemExtent,
+      this.animateMenuItems,
+      this.blurSize,
+      this.blurBackgroundColor,
+      this.menuWidth,
+      this.bottomOffsetHeight,
+      this.menuOffset,
+      this.openWithTap = false})
+      : super(key: key);
+
+  @override
+  _FocusedMenuHolderState createState() => _FocusedMenuHolderState();
+}
+
+class _FocusedMenuHolderState extends State<FocusedMenuHolder> {
+  GlobalKey containerKey = GlobalKey();
+  Offset childOffset = Offset(0, 0);
+  Size? childSize;
+
+  getOffset() {
+    RenderBox renderBox =
+        containerKey.currentContext!.findRenderObject() as RenderBox;
+    Size size = renderBox.size;
+    Offset offset = renderBox.localToGlobal(Offset.zero);
+    setState(() {
+      this.childOffset = Offset(offset.dx, offset.dy);
+      childSize = size;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+        key: containerKey,
+        onTap: () async {
+          widget.onPressed();
+          if (widget.openWithTap) {
+            await openMenu(context);
+          }
+        },
+        onLongPress: () async {
+          if (!widget.openWithTap) {
+            await openMenu(context);
+          }
+        },
+        child: widget.child);
+  }
+
+  Future openMenu(BuildContext context) async {
+    getOffset();
+    await Navigator.push(
+      context,
+      PageRouteBuilder(
+        transitionDuration: widget.duration ?? Duration(milliseconds: 100),
+        pageBuilder: (context, animation, secondaryAnimation) {
+          animation = Tween(begin: 0.0, end: 1.0).animate(animation);
+          return FadeTransition(
+            opacity: animation,
+            child: FocusedMenuDetails(
+              itemExtent: widget.menuItemExtent,
+              menuBoxDecoration: widget.menuBoxDecoration,
+              child: widget.child,
+              childOffset: childOffset,
+              childSize: childSize,
+              menuItems: widget.menuItems,
+              blurSize: widget.blurSize,
+              menuWidth: widget.menuWidth,
+              blurBackgroundColor: widget.blurBackgroundColor,
+              animateMenu: widget.animateMenuItems ?? true,
+              bottomOffsetHeight: widget.bottomOffsetHeight ?? 0,
+              menuOffset: widget.menuOffset ?? 0,
+            ),
+          );
+        },
+        fullscreenDialog: true,
+        opaque: false,
+      ),
+    );
+  }
+}
